+++ jiraone/__init__.py
@@ -22,11 +22,12 @@ and many more depending on what you can come up with.
 """
 from jiraone.access import LOGIN, endpoint, echo, For, field
 from jiraone.jira_logs import add_log, WORK_PATH
-from jiraone.reporting import PROJECT, USER, file_writer, file_reader, path_builder, replacement_placeholder, comment
+from jiraone.reporting import PROJECT, USER, file_writer, file_reader, path_builder, replacement_placeholder, comment, \
+    delete_attachments
 from jiraone.management import manage
 
 __author__ = "Prince Nyeche"
-__version__ = "0.5.6"
+__version__ = "0.5.7"
 __all__ = ["LOGIN",
            "endpoint",
            "echo",
@@ -39,4 +40,4 @@ __all__ = ["LOGIN",
            "path_builder",
            "replacement_placeholder",
            "For",
-           "field", "comment", "manage"]
+           "field", "comment", "manage", "delete_attachments"]